+++ noxfile.py
@@ -74,7 +74,7 @@ def safety(session: nox.Session) -> None:
 @nox.session(name="docs-build", python=python_versions[0])
 def docs_build(session: nox.Session) -> None:
     """Build the documentation."""
-    args = session.posargs or ["docs/source", "docs/build"]
+    args = session.posargs or ["docs/source", "docs/build/html"]
     if not session.posargs and "FORCE_COLOR" in os.environ:
         args.insert(0, "--color")
 
@@ -91,11 +91,11 @@ def docs_build(session: nox.Session) -> None:
 @nox.session(python=python_versions[0])
 def docs(session: nox.Session) -> None:
     """Build and serve the documentation with live reloading on file changes."""
-    args = session.posargs or ["--open-browser", "docs/source", "docs/build"]
+    args = session.posargs or ["--open-browser", "docs/source", "docs/build/html"]
     session.install(".[doc]")
     # session.install("sphinx", "sphinx-autobuild", "sphinx-click", "furo", "myst-parser")
 
-    build_dir = Path("docs", "build")
+    build_dir = Path("docs", "build", "html")
     if build_dir.exists():
         shutil.rmtree(build_dir)