+++ src/logbook_parser/project_uuid.py
@@ -1,3 +1,0 @@
-from uuid import uuid5, NAMESPACE_DNS
-
-PROJECT_UUID = uuid5(NAMESPACE_DNS, "aa_logbook_parser.pfmsoft.com")