+++ CHANGELOG.md
@@ -2,11 +2,34 @@
 <!-- markdownlint-disable MD024 -->
 <!-- changelog-start -->
 
-## [Unreleased](https://github.com/DonalChilde/logbook-parser/v0.0.0...refs/heads/develop)
+## [Unreleased](https://github.com/DonalChilde/logbook-parser/v0.1.0...refs/heads/develop)
 <!-- Dont forget to update the Unreleased compare version to latest release tag -->
 <!-- Copy paste release notes below here -->
 
-## v0.0.0 - 2022-10-3
+## [v0.1.0 2023-04-11](https://github.com/DonalChilde/logbook-parser/compare/v0.0.0...v0.1.0)
+
+### What's Changed in v0.1.0
+
+This is the first release of a functional version of the APA Logbook Parser.
+
+### Added
+
+- Update cli help ([#23](https://github.com/DonalChilde/logbook-parser/pull/23))
+- Add basic stats as `__str__` output ([#21](https://github.com/DonalChilde/logbook-parser/pull/21))
+- Add Default file names ([#18](https://github.com/DonalChilde/logbook-parser/pull/18))
+- isorted imports, mixins for json and csv ([#17](https://github.com/DonalChilde/logbook-parser/pull/17))
+- Expand raw logbook ([#14](https://github.com/DonalChilde/logbook-parser/pull/14))
+
+### Changed
+
+- Parsed metadata ([#13](https://github.com/DonalChilde/logbook-parser/pull/13))
+- Update project to new layout ([#5](https://github.com/DonalChilde/logbook-parser/pull/5))
+- Rebrand as apa logbook parser ([#22](https://github.com/DonalChilde/logbook-parser/pull/22))
+- refactor cli interface - parse xml file ([#12](https://github.com/DonalChilde/logbook-parser/pull/12))
+- 10 use pydantic for data models ([#11](https://github.com/DonalChilde/logbook-parser/pull/11))
+- move snippets code to snippets package ([#7](https://github.com/DonalChilde/logbook-parser/pull/7))
+
+## v0.0.0 - 2022-10-03
 
 ### Whats Changed in v0.0.0